+++ simbolbackend/database/migrations/2017_12_26_225952_add_postura_contraparte_id_to_posturasMatches.php
@@ -36,9 +36,12 @@ class AddPosturaContraparteIdToPosturasMatches extends Migration
      */
     public function down()
     {
+      if (Schema::hasTable('posturasMatches'))
+      {
         Schema::table('posturasMatches', function (Blueprint $table) {
-            //
+            $table->dropForeign(['postura_contraparte_id']);
             $table->dropColumn('postura_contraparte_id');
         });
+      }
     }
 }
