+++ simbolbackend/database/migrations/2018_02_15_142905_add_column_id_status_notifications_to_notificacions.php
@@ -13,13 +13,15 @@ class AddColumnIdStatusNotificationsToNotificacions extends Migration
      */
     public function up()
     {
-        //
+      if(!Schema::hasColumn('notificacions', 'status_notifications_id'))
+      {
         Schema::table('notificacions', function (Blueprint $table) {
             //
             $table->integer('status_notifications_id')->unsigned()->default(1);
             
             $table->foreign('status_notifications_id')->references('id')->on('status_notifications')->onDelete('cascade');
         });
+      }
     }
 
     /**
@@ -29,10 +31,9 @@ class AddColumnIdStatusNotificationsToNotificacions extends Migration
      */
     public function down()
     {
-        //
-        Schema::table('notificacions', function (Blueprint $table) {
-            //
-            $table->dropColumn('status_notifications_id');
-        });
+      Schema::table('notificacions', function (Blueprint $table) {
+        $table->dropForeign(['status_notifications_id']);
+        $table->dropColumn('status_notifications_id');
+      });
     }
 }
