+++ simbolbackend/database/migrations/2018_03_12_135725_add_column_posturas_match_id_to_log_users.php
@@ -13,13 +13,15 @@ class AddColumnPosturasMatchIdToLogUsers extends Migration
      */
     public function up()
     {
-        //
-        Schema::table('log_users', function (Blueprint $table) {
-            //
-            $table->integer('posturas_match_id')->default(0)->unsigned();
+      Schema::table('log_users', function (Blueprint $table)
+      {
+        if(!Schema::hasColumn('log_users', 'posturas_match_id'))
+        {
+          $table->integer('posturas_match_id')->default(0)->unsigned();
 
-            $table->foreign('posturas_match_id')->references('idposturasMatch')->on('posturasMatches')->onDelete('cascade');
-        });
+          $table->foreign('posturas_match_id')->references('idposturasMatch')->on('posturasMatches')->onDelete('cascade');
+        }
+      });
     }
 
     /**
@@ -30,9 +32,12 @@ class AddColumnPosturasMatchIdToLogUsers extends Migration
     public function down()
     {
         //
-        Schema::table('log_users', function (Blueprint $table) {
-            //
+        if(Schema::hasColumn('log_users', 'posturas_match_id'))
+        {
+          Schema::table('log_users', function (Blueprint $table) {
+            $table->dropForeign(['posturas_match_id']);
             $table->dropColumn('posturas_match_id');
-        });
+          });
+        }
     }
 }
